add imx8 smmu stream id initialization
This setups the stream id table for the masters used in the bao
linux+freertos demo. The reason for this is that setting these is is
very system configuration dependent, varying given the number of guests
and how these dma-capable peripherals are assigned to each one.

Ideally this would be done directly by bao through some kind of platform
specific iommu_set_id function. Possibly this function could just call
monitor could so that we don't have to add scu communication code to the
hypervisor.