feat(build): allow atf to build partitioner sources with equal filenames
Add to the partitioner.mk a method to deal with source files that have
the same filename in the partitioner module (e.g. arch/vmm and core/vmm).

The temporary copy of the files is made and passed to the BL31_SOURCES
variable.

The clean target is also modified with a internal targett dependency to
eliminate the temporary files.

Signed-off-by: Daniel Oliveira <drawnpoetry@gmail.com>